Move inner classes to standalone classes (#1095)
