+++ polaris-core/src/main/java/org/apache/polaris/core/persistence/LocalPolarisMetaStoreManagerFactory.java
@@ -36,6 +36,8 @@ import org.apache.polaris.core.entity.PolarisEntityType;
 import org.apache.polaris.core.entity.PolarisPrincipalSecrets;
 import org.apache.polaris.core.persistence.bootstrap.RootCredentialsSet;
 import org.apache.polaris.core.persistence.cache.EntityCache;
+import org.apache.polaris.core.persistence.dao.entity.BaseResult;
+import org.apache.polaris.core.persistence.dao.entity.EntityResult;
 import org.apache.polaris.core.persistence.transactional.PolarisMetaStoreManagerImpl;
 import org.apache.polaris.core.persistence.transactional.TransactionalPersistence;
 import org.apache.polaris.core.storage.cache.StorageCredentialCache;
@@ -198,7 +200,7 @@ public abstract class LocalPolarisMetaStoreManagerFactory<StoreType>
       CallContext.setCurrentContext(CallContext.of(realmContext, polarisContext));
     }
 
-    PolarisMetaStoreManager.EntityResult preliminaryRootPrincipalLookup =
+    EntityResult preliminaryRootPrincipalLookup =
         metaStoreManager.readEntityByName(
             polarisContext,
             null,
@@ -215,7 +217,7 @@ public abstract class LocalPolarisMetaStoreManagerFactory<StoreType>
 
     metaStoreManager.bootstrapPolarisService(polarisContext);
 
-    PolarisMetaStoreManager.EntityResult rootPrincipalLookup =
+    EntityResult rootPrincipalLookup =
         metaStoreManager.readEntityByName(
             polarisContext,
             null,
@@ -256,7 +258,7 @@ public abstract class LocalPolarisMetaStoreManagerFactory<StoreType>
       CallContext.setCurrentContext(CallContext.of(realmContext, polarisContext));
     }
 
-    PolarisMetaStoreManager.EntityResult rootPrincipalLookup =
+    EntityResult rootPrincipalLookup =
         metaStoreManager.readEntityByName(
             polarisContext,
             null,
