+++ polaris-core/src/main/java/org/apache/polaris/core/persistence/dao/entity/EntityWithPath.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.polaris.core.persistence.dao.entity;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.annotation.Nonnull;
+import java.util.List;
+import org.apache.polaris.core.entity.PolarisBaseEntity;
+import org.apache.polaris.core.entity.PolarisEntityCore;
+
+/** Class to represent an entity with its path */
+public class EntityWithPath {
+  // path to that entity. Could be null if this entity is top-level
+  private final @Nonnull List<PolarisEntityCore> catalogPath;
+
+  // the base entity itself
+  private final @Nonnull PolarisBaseEntity entity;
+
+  @JsonCreator
+  public EntityWithPath(
+      @JsonProperty("catalogPath") @Nonnull List<PolarisEntityCore> catalogPath,
+      @JsonProperty("entity") @Nonnull PolarisBaseEntity entity) {
+    this.catalogPath = catalogPath;
+    this.entity = entity;
+  }
+
+  public @Nonnull List<PolarisEntityCore> getCatalogPath() {
+    return catalogPath;
+  }
+
+  public @Nonnull PolarisBaseEntity getEntity() {
+    return entity;
+  }
+}
