+++ quarkus/service/src/test/java/org/apache/polaris/service/quarkus/catalog/PolarisCatalogHandlerWrapperAuthzTest.java
@@ -62,7 +62,7 @@ import org.apache.polaris.core.entity.CatalogRoleEntity;
 import org.apache.polaris.core.entity.PolarisPrivilege;
 import org.apache.polaris.core.entity.PrincipalEntity;
 import org.apache.polaris.core.persistence.PolarisEntityManager;
-import org.apache.polaris.core.persistence.PolarisMetaStoreManager;
+import org.apache.polaris.core.persistence.dao.entity.CreatePrincipalResult;
 import org.apache.polaris.core.persistence.resolver.PolarisResolutionManifest;
 import org.apache.polaris.service.catalog.PolarisCatalogHandlerWrapper;
 import org.apache.polaris.service.catalog.io.DefaultFileIOFactory;
@@ -227,7 +227,7 @@ public class PolarisCatalogHandlerWrapperAuthzTest extends PolarisAuthzTestBase 
   @Test
   public void testInsufficientPermissionsPriorToSecretRotation() {
     String principalName = "all_the_powers";
-    PolarisMetaStoreManager.CreatePrincipalResult newPrincipal =
+    CreatePrincipalResult newPrincipal =
         metaStoreManager.createPrincipal(
             callContext.getPolarisCallContext(),
             new PrincipalEntity.Builder()
