+++ service/common/src/main/java/org/apache/polaris/service/auth/TokenBroker.java
@@ -25,6 +25,7 @@ import org.apache.polaris.core.auth.PolarisSecretsManager.PrincipalSecretsResult
 import org.apache.polaris.core.entity.PolarisEntityType;
 import org.apache.polaris.core.entity.PrincipalEntity;
 import org.apache.polaris.core.persistence.PolarisMetaStoreManager;
+import org.apache.polaris.core.persistence.dao.entity.EntityResult;
 import org.apache.polaris.service.types.TokenType;
 
 /** Generic token class intended to be extended by different token types */
@@ -124,7 +125,7 @@ public interface TokenBroker {
     if (!principalSecrets.getPrincipalSecrets().matchesSecret(clientSecret)) {
       return Optional.empty();
     }
-    PolarisMetaStoreManager.EntityResult result =
+    EntityResult result =
         metaStoreManager.loadEntity(
             polarisCallContext, 0L, principalSecrets.getPrincipalSecrets().getPrincipalId());
     if (!result.isSuccess() || result.getEntity().getType() != PolarisEntityType.PRINCIPAL) {
