+++ service/common/src/main/java/org/apache/polaris/service/catalog/BasePolarisCatalog.java
@@ -89,11 +89,14 @@ import org.apache.polaris.core.entity.PolarisEntitySubType;
 import org.apache.polaris.core.entity.PolarisEntityType;
 import org.apache.polaris.core.entity.PolarisTaskConstants;
 import org.apache.polaris.core.entity.TableLikeEntity;
-import org.apache.polaris.core.persistence.BaseResult;
 import org.apache.polaris.core.persistence.PolarisEntityManager;
 import org.apache.polaris.core.persistence.PolarisMetaStoreManager;
 import org.apache.polaris.core.persistence.PolarisResolvedPathWrapper;
 import org.apache.polaris.core.persistence.ResolvedPolarisEntity;
+import org.apache.polaris.core.persistence.dao.entity.BaseResult;
+import org.apache.polaris.core.persistence.dao.entity.DropEntityResult;
+import org.apache.polaris.core.persistence.dao.entity.EntityResult;
+import org.apache.polaris.core.persistence.dao.entity.ListEntitiesResult;
 import org.apache.polaris.core.persistence.resolver.PolarisResolutionManifest;
 import org.apache.polaris.core.persistence.resolver.PolarisResolutionManifestCatalogView;
 import org.apache.polaris.core.persistence.resolver.ResolverPath;
@@ -432,7 +435,7 @@ public class BasePolarisCatalog extends BaseMetastoreViewCatalog
                   return clone;
                 })
             .orElse(Map.of());
-    PolarisMetaStoreManager.DropEntityResult dropEntityResult =
+    DropEntityResult dropEntityResult =
         dropTableLike(PolarisEntitySubType.TABLE, tableIdentifier, storageProperties, purge);
     if (!dropEntityResult.isSuccess()) {
       return false;
@@ -637,7 +640,7 @@ public class BasePolarisCatalog extends BaseMetastoreViewCatalog
 
     // drop if exists and is empty
     PolarisCallContext polarisCallContext = callContext.getPolarisCallContext();
-    PolarisMetaStoreManager.DropEntityResult dropEntityResult =
+    DropEntityResult dropEntityResult =
         getMetaStoreManager()
             .dropEntityIfExists(
                 getCurrentPolarisContext(),
@@ -1077,7 +1080,7 @@ public class BasePolarisCatalog extends BaseMetastoreViewCatalog
    */
   private void validateNoLocationOverlap(
       String location, List<PolarisEntity> parentPath, String name) {
-    PolarisMetaStoreManager.ListEntitiesResult siblingNamespacesResult =
+    ListEntitiesResult siblingNamespacesResult =
         getMetaStoreManager()
             .listEntities(
                 callContext.getPolarisCallContext(),
@@ -1100,7 +1103,7 @@ public class BasePolarisCatalog extends BaseMetastoreViewCatalog
         parentNamespace
             .map(
                 ns -> {
-                  PolarisMetaStoreManager.ListEntitiesResult siblingTablesResult =
+                  ListEntitiesResult siblingTablesResult =
                       getMetaStoreManager()
                           .listEntities(
                               callContext.getPolarisCallContext(),
@@ -1690,7 +1693,7 @@ public class BasePolarisCatalog extends BaseMetastoreViewCatalog
     }
 
     // rename the entity now
-    PolarisMetaStoreManager.EntityResult returnedEntityResult =
+    EntityResult returnedEntityResult =
         getMetaStoreManager()
             .renameEntity(
                 getCurrentPolarisContext(),
@@ -1860,7 +1863,7 @@ public class BasePolarisCatalog extends BaseMetastoreViewCatalog
   }
 
   @SuppressWarnings("FormatStringAnnotation")
-  private @Nonnull PolarisMetaStoreManager.DropEntityResult dropTableLike(
+  private @Nonnull DropEntityResult dropTableLike(
       PolarisEntitySubType subType,
       TableIdentifier identifier,
       Map<String, String> storageProperties,
@@ -1869,8 +1872,7 @@ public class BasePolarisCatalog extends BaseMetastoreViewCatalog
         resolvedEntityView.getResolvedPath(identifier, subType);
     if (resolvedEntities == null) {
       // TODO: Error?
-      return new PolarisMetaStoreManager.DropEntityResult(
-          BaseResult.ReturnStatus.ENTITY_NOT_FOUND, null);
+      return new DropEntityResult(BaseResult.ReturnStatus.ENTITY_NOT_FOUND, null);
     }
 
     List<PolarisEntity> catalogPath = resolvedEntities.getRawParentPath();
